minimized reusing of functions/variables
