+++ src/pages/blocks/[blockId]/index.page.tsx
@@ -16,11 +16,12 @@ import { Head } from "@components/commons/Head";
 import { Container } from "@components/commons/Container";
 import { MetascanLinkButton } from "@components/commons/MetascanLinkButton";
 import { useState } from "react";
-import { EnvironmentNetwork } from "@waveshq/walletkit-core";
 import { useNetwork } from "@contexts/NetworkContext";
 import { isAlphanumeric } from "../../../utils/commons/StringValidator";
 import { BlockTransactions } from "./_components/BlockTransactions";
 import { BlockDetailTable } from "./_components/BlockDetailTable";
+import { VmmapResult, VmmapTypes } from "../../transactions/enum/VmmapTypes";
+import { getMetaScanBlockUrl } from "../../../utils/commons/getNetworkParams";
 
 interface BlockDetailsPageProps {
   block: Block;
@@ -29,15 +30,6 @@ interface BlockDetailsPageProps {
     pages: CursorPage[];
   };
   mappedEvmBlockNum?: string | null;
-}
-interface VmmapResult {
-  input: string;
-  type: string;
-  output: string;
-}
-
-enum VmmapTypes {
-  BlockNumberDVMToEVM = 1,
 }
 
 export default function BlockDetails(
@@ -102,35 +94,6 @@ function BlockHeading({
       </div>
     </div>
   );
-}
-
-function getNetworkParams(network: EnvironmentNetwork): string {
-  switch (network) {
-    case EnvironmentNetwork.MainNet:
-      // no-op: network param not required for MainNet
-      return "";
-    case EnvironmentNetwork.TestNet:
-      return `?network=${EnvironmentNetwork.TestNet}`;
-    case EnvironmentNetwork.DevNet:
-      return `?network=${EnvironmentNetwork.DevNet}`;
-
-    case EnvironmentNetwork.LocalPlayground:
-    case EnvironmentNetwork.RemotePlayground:
-      return `?network=${EnvironmentNetwork.RemotePlayground}`;
-    case EnvironmentNetwork.Changi:
-      return `?network=${EnvironmentNetwork.Changi}`;
-    default:
-      return "";
-  }
-}
-
-function getMetaScanBlockUrl(
-  network: EnvironmentNetwork,
-  id?: string | null,
-): string {
-  const baseMetaScanUrl = "https://meta.defiscan.live";
-  const networkParams = getNetworkParams(network);
-  return `${baseMetaScanUrl}/block/${id}${networkParams}`;
 }
 
 export async function getServerSideProps(
